+++ upnp/src/inc/upnputil.h
@@ -5,6 +5,7 @@
  *
  * Copyright (c) 2000-2003 Intel Corporation
  * All rights reserved.
+ * Copyright (c) 2012 France Telecom All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -109,7 +110,7 @@ void linecopylen(
 #endif
 
 /* Size of the errorBuffer variable, passed to the strerror_r() function */
-#define ERROR_BUFFER_LEN 256
+#define ERROR_BUFFER_LEN (size_t)256
 
 /* C specific */
 /* VC needs these in C++ mode too (do other compilers?) */
